added adaptiveSubdivision to halfedgetools
it took a few hrs work to make gonska's subdivider parametrized. it is still no optimial, but in principle it no longer depends in SubDivEdge etc. the only criteria is HasLengthSquared. this should probably be merged into the delaunay predicate HasLength.

this brings into light that JREdge should have getLength() since it corresponds to and IFS and jReality can calc. distances. however, then we need the signature of the metric. this should be stored in something like JRHDS.

this version needs cleanup, but in the same way HalfedgeIO and Debug need it (casts to types etc.) we should discuss the proper way to parametrize plugins: basically hedsconnector is parametrized, so everything else needs to be it to?

ideas?
kristoffer

git-svn-id: https://svn.math.tu-berlin.de/svn/jtem@465 f5b180c5-49ee-4939-b20e-b6ed35f0f7b7